+++ app/src/main/java/com/vultisig/wallet/ui/models/WelcomeViewModel.kt
@@ -5,6 +5,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.vultisig.wallet.R
 import com.vultisig.wallet.data.models.OnBoardPage
 import com.vultisig.wallet.data.repositories.OnBoardRepository
 import com.vultisig.wallet.data.repositories.VaultRepository
@@ -39,7 +40,7 @@ internal class WelcomeViewModel @Inject constructor(
     var channel = _channel.receiveAsFlow()
 
     init {
-        getBoardPages()
+        state = WelcomeState(pages)
     }
 
     fun scrollToNextPage() {
@@ -56,10 +57,6 @@ internal class WelcomeViewModel @Inject constructor(
         saveOnBoardingState()
     }
 
-    private fun getBoardPages() {
-        state = state.copy(pages = repository.onBoardPages())
-    }
-
     private fun saveOnBoardingState() {
         viewModelScope.launch(Dispatchers.IO) {
             repository.saveOnBoardingState(completed = true)
@@ -71,5 +68,24 @@ internal class WelcomeViewModel @Inject constructor(
             navigator.navigate(dest)
         }
     }
-
 }
+
+private val pages = listOf(
+    OnBoardPage(
+        image = R.drawable.intro1,
+        title = R.string.onboard_intro1_title,
+        description = R.string.onboard_intro1_desc
+    ), OnBoardPage(
+        image = R.drawable.intro2,
+        title = R.string.onboard_intro2_title,
+        description = R.string.onboard_intro2_desc,
+    ), OnBoardPage(
+        image = R.drawable.intro3,
+        title = R.string.onboard_intro3_title,
+        description = R.string.onboard_intro3_desc,
+    ), OnBoardPage(
+        image = R.drawable.intro4,
+        title = R.string.onboard_intro4_title,
+        description = R.string.onboard_intro4_desc,
+    )
+)
